MagNav: airborne Magnetic anomaly Navigation

Placeholder page for MagNav.jl documentation.

MagNav.create_TL_A — Function

create_TL_A(Bx, By, Bz;
            Bt       = sqrt.(Bx.^2+By.^2+Bz.^2),
            terms    = [:permanent,:induced,:eddy],
            Bt_scale = 50000)

Create Tolles-Lawson A matrix using vector magnetometer measurements.

Arguments:

Bx,By,Bz: vector magnetometer measurements [nT]

Bt: (optional) magnitude of vector magnetometer measurements or scalar magnetometer measurements for modified Tolles-Lawson [nT]

terms: (optional) Tolles-Lawson terms to use {:permanent,:induced,:eddy,:bias}

Bt_scale: (optional) scaling factor for induced and eddy current terms [nT]

Returns:

A: Tolles-Lawson A matrix

create_TL_A(flux::MagV, ind=trues(length(flux.x));
            Bt       = sqrt.(flux.x.^2+flux.y.^2+flux.z.^2)[ind],
            terms    = [:permanent,:induced,:eddy],
            Bt_scale = 50000)

Create Tolles-Lawson A matrix using vector magnetometer measurements.

Arguments:

flux: MagV vector magnetometer measurement struct

ind: (optional) selected data indices

Bt: (optional) magnitude of vector magnetometer measurements or scalar magnetometer measurements for modified Tolles-Lawson [nT]

terms: (optional) Tolles-Lawson terms to use {:permanent,:induced,:eddy,:bias}

Bt_scale: (optional) scaling factor for induced and eddy current terms [nT]

Returns:

A: Tolles-Lawson A matrix

MagNav.create_TL_coef — Function

create_TL_coef(Bx, By, Bz, B;
               Bt         = sqrt.(Bx.^2+By.^2+Bz.^2),
               λ          = 0,
               terms      = [:permanent,:induced,:eddy],
               pass1      = 0.1,
               pass2      = 0.9,
               fs         = 10.0,
               pole::Int  = 4,
               trim::Int  = 20,
               Bt_scale   = 50000,
               return_var = false)

Create Tolles-Lawson coefficients using vector and scalar magnetometer measurements and a bandpass, low-pass or high-pass filter.

Arguments:

Bx,By,Bz: vector magnetometer measurements [nT]

B: scalar magnetometer measurements [nT]

Bt: (optional) magnitude of vector magnetometer measurements or scalar magnetometer measurements for modified Tolles-Lawson [nT]

λ: (optional) ridge parameter

terms: (optional) Tolles-Lawson terms to use {:permanent,:induced,:eddy,:bias}

pass1: (optional) first passband frequency [Hz]

pass2: (optional) second passband frequency [Hz]

fs: (optional) sampling frequency [Hz]

pole: (optional) number of poles for Butterworth filter

trim (optional) number of elements to trim after filtering

Bt_scale: (optional) scaling factor for induced and eddy current terms [nT]

return_var: (optional) return B_var fit error variance along with coef

Returns:

coef: Tolles-Lawson coefficients

B_var: if return_var=true, also return fit error variance

create_TL_coef(flux::MagV, B, ind=trues(length(flux.x));
               Bt         = sqrt.(flux.x.^2+flux.y.^2+flux.z.^2)[ind],
               λ          = 0,
               terms      = [:permanent,:induced,:eddy],
               pass1      = 0.1,
               pass2      = 0.9,
               fs         = 10.0,
               pole::Int  = 4,
               trim::Int  = 20,
               Bt_scale   = 50000,
               return_var = false)

Create Tolles-Lawson coefficients using vector and scalar magnetometer measurements and a bandpass, low-pass or high-pass filter.

Arguments:

flux: MagV vector magnetometer measurement struct

B: scalar magnetometer measurements [nT]

ind: (optional) selected data indices

Bt: (optional) magnitude of vector magnetometer measurements or scalar magnetometer measurements for modified Tolles-Lawson [nT]

λ: (optional) ridge parameter

terms: (optional) Tolles-Lawson terms to use {:permanent,:induced,:eddy,:bias}

pass1: (optional) first passband frequency [Hz]

pass2: (optional) second passband frequency [Hz]

fs: (optional) sampling frequency [Hz]

pole: (optional) number of poles for Butterworth filter

trim (optional) number of elements to trim after filtering

Bt_scale: (optional) scaling factor for induced and eddy current terms [nT]

return_var: (optional) return B_var fit error variance along with coef

Returns:

coef: Tolles-Lawson coefficients

B_var: fit error variance

MagNav.fdm — Function

fdm(x::Vector; central::Bool=true, fourth::Bool=false)

Finite difference method (FDM) on vector of input data.

Arguments:

x: input data

central: (optional) if true, 1st-order central difference, otherwise backward

fourth: (optional) if true, 4th-order central difference

Returns:

dif: length of x finite differences

Version

julia> VERSIONv"1.7.3"