Magnetic Anomaly Maps

The following are key functions related to magnetic anomaly maps.

Load Map Data

MagNav.get_map — Function

get_map(map_file::String; map_units::Symbol=:deg)

Get map data from saved file. Map files are typically saved with :deg units.

Arguments:

map_file: path/name of magnetic anomaly map HDF5 or MAT file

map_units: (optional) map xx/yy units used in HDF5 file {:deg,:rad}

Returns:

mapS or mapV: MapS scalar or MapV vector magnetic anomaly map struct

get_map(map_name::Symbol, df_map::DataFrame; map_units::Symbol=:deg)

Get map data from saved file via DataFrame lookup. Map files are typically saved with :deg units.

Arguments:

map_name: name of magnetic anomaly map

df_map: lookup table (DataFrame) of map files

map_units: (optional) map xx/yy units used in HDF5 file {:deg,:rad}

Returns:

mapS or mapV: MapS scalar or MapV vector magnetic anomaly map struct

Upward Continue Map

MagNav.upward_fft — Function

upward_fft(map_map, dx, dy, dz; expand::Bool=true, α=0)

Upward continuation of a potential field (i.e. magnetic anomaly field) map. Uses the Fast Fourier Transform to convert the map to the frequency domain, applies an upward continuation filter, and converts back to the spatial domain. Optionally expands the map temporarily with periodic padding. Downward continuation can be done to a limited degree as well, but be careful as this can be unstable and amplify high frequencies (i.e. noise).

Reference: Blakely, Potential Theory in Gravity and Magnetic Applications, 2009, Chapter 12 & Appendix B (pg. 315-317 & 402).

Arguments

map_map: ny x nx 2D gridded map data

dx: x-direction map sample interval [m]

dy: y-direction map sample interval [m]

dz: z-direction upward continuation distance [m]

expand: (optional) if true, expand map temporarily to limit edge effects

α: (optional) regularization parameter for downward continuation

Returns

map_out: ny x nx upward continued 2D gridded map data

upward_fft(map_map::Union{MapS,MapV}, alt; expand::Bool=true, α=0)

Arguments

map_map: MapS scalar or MapV vector magnetic anomaly map struct

alt: target upward continuation altitude [m]

expand: (optional) if true, expand map temporarily to limit edge effects

α: (optional) regularization parameter for downward continuation

Returns

map_out: MapS scalar or MapV vector magnetic anomaly map struct with upward continued map

Interpolate Map

MagNav.map_interpolate — Function

map_interpolate(map_map, map_xx, map_yy, type::Symbol = :cubic)

Create map grid interpolation, equivalent of griddedInterpolant in MATLAB. Uses the Interpolations package rather than Dierckx or GridInterpolations, as Interpolations was found to be fastest for MagNav use cases.

Arguments:

map_map: ny x nx 2D gridded map data

map_xx: nx x-direction map coordinates

map_yy: ny y-direction map coordinates

type: (optional) type of interpolation {:linear,:quad,:cubic}

Returns:

itp_map: map grid interpolation

map_interpolate(mapS::MapS, type::Symbol = :cubic; vert::Bool = false)

Create map grid interpolation, equivalent of griddedInterpolant in MATLAB. Optionally return vertical derivative grid interpolation, which is calculated using finite differences between the map and a slightly upward continued map.

Arguments:

mapS: MapS scalar magnetic anomaly map struct

type: (optional) type of interpolation {:linear,:quad,:cubic}

vert: (optional) if true, also return vertical derivative grid interpolation

Returns:

itp_map: map grid interpolation

map_interpolate(mapV::MapV, dim::Symbol = :X, type::Symbol = :cubic)

Create map grid interpolation, equivalent of griddedInterpolant in MATLAB.

Arguments:

mapV: MapV vector magnetic anomaly map struct

dim: map dimension to interpolate {:X,:Y,:Z}

type: (optional) type of interpolation {:linear,:quad,:cubic}

Returns:

itp_map: map grid interpolation

Plot Map

MagNav.plot_map — Function

plot_map(map_map::Matrix,
         map_xx::Vector     = [],
         map_yy::Vector     = [];
         clims::Tuple       = (0,0),
         dpi::Int           = 200,
         margin::Int        = 2,
         legend::Bool       = true,
         axis::Bool         = true,
         fewer_pts::Bool    = true,
         map_color::Symbol  = :usgs,
         bg_color::Symbol   = :white,
         map_units::Symbol  = :rad,
         plot_units::Symbol = :deg,
         b_e                = gr())

Plot map.

Arguments:

map_map: ny x nx 2D gridded map data

map_xx: nx x-direction (longitude) map coordinates [rad] or [deg]

map_yy: ny y-direction (latitude) map coordinates [rad] or [deg]

clims: (optional) color scale limits

dpi: (optional) dots per inch (image resolution)

margin: (optional) margin around plot [mm]

legend: (optional) if true, show legend

axis: (optional) if true, show axes

fewer_pts: (optional) if true, reduce number of data points plotted

map_color: (optional) filled contour color scheme {:usgs,:gray,:gray1,:gray2,:plasma,:magma}

bg_color: (optional) background color

map_units: (optional) map xx/yy units {:rad,:deg}

plot_units: (optional) plot xx/yy units {:rad,:deg,:m}

b_e: (optional) plotting backend

Returns:

p1: plot with map

plot_map(mapS::Union{MapS,MapSd};
         clims::Tuple       = (0,0),
         dpi::Int           = 200,
         margin::Int        = 2,
         legend::Bool       = true,
         axis::Bool         = true,
         fewer_pts::Bool    = true,
         map_color::Symbol  = :usgs,
         bg_color::Symbol   = :white,
         map_units::Symbol  = :rad,
         plot_units::Symbol = :deg,
         b_e                = gr())

Plot map.

Arguments:

mapS: MapS or MapSd scalar magnetic anomaly map struct

clims: (optional) color scale limits

dpi: (optional) dots per inch (image resolution)

margin: (optional) margin around plot [mm]

legend: (optional) if true, show legend

axis: (optional) if true, show axes

fewer_pts: (optional) if true, reduce number of data points plotted

map_color: (optional) filled contour color scheme {:usgs,:gray,:gray1,:gray2,:plasma,:magma}

bg_color: (optional) background color

map_units: (optional) map xx/yy units {:rad,:deg}

plot_units: (optional) plot xx/yy units {:rad,:deg,:m}

b_e: (optional) plotting backend

Returns:

p1: plot with map