Custom Structs

The following are custom structs used within the package.

Magnetic Anomaly Maps

Missing docstring.

Missing docstring for Map. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MapS. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MapSd. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MapV. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MapVd. Check Documenter's build log for details.

Vector Magnetometer

Missing docstring.

Missing docstring for MagV. Check Documenter's build log for details.

Flight Paths

Missing docstring.

Missing docstring for Path. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Traj. Check Documenter's build log for details.

Missing docstring.

Missing docstring for INS. Check Documenter's build log for details.

Flight Data

Missing docstring.

Missing docstring for XYZ. Check Documenter's build log for details.

Missing docstring.

Missing docstring for XYZ0. Check Documenter's build log for details.

Missing docstring.

Missing docstring for XYZ1. Check Documenter's build log for details.

Missing docstring.

Missing docstring for XYZ20. Check Documenter's build log for details.

Missing docstring.

Missing docstring for XYZ21. Check Documenter's build log for details.

Output Wrappers

Missing docstring.

Missing docstring for FILTres. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CRLBout. Check Documenter's build log for details.

Missing docstring.

Missing docstring for INSout. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FILTout. Check Documenter's build log for details.

Aeromagnetic Compensation Model Parameters

Missing docstring.

Missing docstring for CompParams. Check Documenter's build log for details.

MagNav.LinCompParams — Type

struct LinCompParams <: CompParams

Linear aeromagnetic compensation parameters struct. Subtype of CompParams.

To see default parameters, type LinCompParams().

General Parameters:

Parameter	Description

features_setup	list of features to include

features_no_norm	list of features to not normalize

model_type	aeromagnetic compensation model type (see below)

y_type	y target type (see below)

use_mag	scalar magnetometer to use {:mag_1_uc, etc.}, only used for y_type = :c, :d, :e

use_vec	vector magnetometer (fluxgate) to use for Tolles-Lawson A matrix {:flux_a, etc.}, not used for model_type = :elasticnet, :plsr

data_norms	Tuple of data normalizations, e.g. (A_bias,A_scale,x_bias,x_scale,y_bias,y_scale)

weights	linear model coefficients

terms	Tolles-Lawson terms to use for Tolles-Lawson A matrix (or matrices) within x matrix {:permanent,:induced,:eddy}

terms_A	Tolles-Lawson terms to use for "external" Tolles-Lawson A matrix {:permanent,:induced,:eddy,:bias}, not used for model_type = :elasticnet, :plsr

sub_diurnal	if true, subtract diurnal from scalar magnetometer measurements

sub_igrf	if true, subtract IGRF from scalar magnetometer measurements

bpf_mag	if true, bpf scalar magnetometer measurements in x matrix

norm_type_A	normalization for Tolles-Lawson A matrix, not used for model_type = :elasticnet, :plsr (see below)

norm_type_x	normalization for x matrix (see below)

norm_type_y	normalization for y target vector (see below)

model_type options:

:TL = classical Tolles-Lawson

:mod_TL = modified Tolles-Lawson

:map_TL = map-based Tolles-Lawson

:elasticnet = elastic net (ridge regression and/or Lasso)

:plsr: = partial least squares regression (PLSR)

y_type options:

:a = anomaly field #1, compensated tail stinger total field scalar magnetometer measurements

:b = anomaly field #2, interpolated magnetic anomaly map values

:c = aircraft field #1, difference between uncompensated cabin total field scalar magnetometer measurements and interpolated magnetic anomaly map values

:d = aircraft field #2, difference between uncompensated cabin and compensated tail stinger total field scalar magnetometer measurements

:e = BPF'd total field, bandpass filtered uncompensated cabin total field scalar magnetometer measurements

norm_type options:

:standardize = Z-score normalization

:normalize = min-max normalization

:scale = scale by maximum absolute value, bias = 0

:none = scale by 1, bias = 0

Linear Model-Specific Parameters:

Parameter	Description

k_plsr	number of components, only used for model_type = :plsr

λ_TL	ridge parameter, only used for model_type = :TL, :mod_TL, :map_TL

MagNav.NNCompParams — Type

struct NNCompParams <: CompParams

Neural network-based aeromagnetic compensation parameters struct. Subtype of CompParams.

To see default parameters, type NNCompParams().

General Parameters:

Parameter	Description

features_setup	list of features to include

features_no_norm	list of features to not normalize

model_type	aeromagnetic compensation model type (see below)

y_type	y target type (see below)

use_mag	scalar magnetometer to use {:mag_1_uc, etc.}, only used for y_type = :c, :d, :e

use_vec	vector magnetometer (fluxgate) to use for Tolles-Lawson A matrix {:flux_a, etc.}, not used for model_type = :m1

data_norms	Tuple of data normalizations, e.g. (A_bias,A_scale,v_scale,x_bias,x_scale,y_bias,y_scale)

weights	neural network model weights

terms	Tolles-Lawson terms to use for Tolles-Lawson A matrix (or matrices) within x matrix {:permanent,:induced,:eddy}

terms_A	Tolles-Lawson terms to use for "external" Tolles-Lawson A matrix {:permanent,:induced,:eddy,:bias}, not used for model_type = :m1

sub_diurnal	if true, subtract diurnal from scalar magnetometer measurements

sub_igrf	if true, subtract IGRF from scalar magnetometer measurements

bpf_mag	if true, bpf scalar magnetometer measurements in x matrix

norm_type_A	normalization for Tolles-Lawson A matrix, not used for model_type = :m1 (see below)

norm_type_x	normalization for x matrix (see below)

norm_type_y	normalization for y target vector (see below)

model_type options:

:m1 = standard NN

:m2a = NN determines Tolles-Lawson coefficients

:m2b = NN determines additive correction to classical Tolles-Lawson

:m2c = NN determines additive correction to classical Tolles-Lawson, Tolles-Lawson coefficients tuned as well

:m2d = NN determines additive correction to each Tolles-Lawson coefficient

y_type options:

:a = anomaly field #1, compensated tail stinger total field scalar magnetometer measurements

:b = anomaly field #2, interpolated magnetic anomaly map values

:c = aircraft field #1, difference between uncompensated cabin total field scalar magnetometer measurements and interpolated magnetic anomaly map values

:d = aircraft field #2, difference between uncompensated cabin and compensated tail stinger total field scalar magnetometer measurements

:e = BPF'd total field, bandpass filtered uncompensated cabin total field scalar magnetometer measurements

norm_type options:

:standardize = Z-score normalization

:normalize = min-max normalization

:scale = scale by maximum absolute value, bias = 0

:none = scale by 1, bias = 0

Neural Network-Based Model-Specific Parameters:

Parameter	Description

TL_coef	Tolles-Lawson coefficients, only used for model_type = :m2b, :m2d (and :m2c initial TL_coef)

η_adam	learning rate for Adam optimizer

epoch_adam	number of epochs for Adam optimizer

epoch_lbfgs	number of epochs for LBFGS optimizer

hidden	hidden layers & nodes, e.g. [8,8] for 2 hidden layers, 8 nodes each

activation	activation function (see below)

batchsize	mini-batch size

frac_train	fraction of training data used for training (remainder for validation), only used for Adam optimizer

α_sgl	Lasso (α_sgl=0) vs group Lasso (α_sgl=1) balancing parameter {0:1}

λ_sgl	sparse group Lasso parameter, typically ~1e-5 (if nonzero)

k_pca	number of components for pre-processing with PCA + whitening, -1 to ignore

drop_fi	if true, perform drop-column feature importance

drop_fi_bson	file name (without extension) to save/load drop-column feature importance data

drop_fi_csv	file name to save drop-column feature importance data

perm_fi	if true, perform permutation feature importance

perm_fi_csv	file name to save permutation feature importance data

activation options:

relu = rectified linear unit

σ = sigmoid (logistic function)

swish = self-gated

tanh = hyperbolic tan

run plot_activation() for a visual