Aeromagnetic Compensation

The following are key functions related to aeromagnetic compensation.

Tolles-Lawson

MagNav.create_TL_A — Function

create_TL_A(Bx, By, Bz;
            Bt       = sqrt.(Bx.^2+By.^2+Bz.^2),
            terms    = [:permanent,:induced,:eddy],
            Bt_scale = 50000)

Create Tolles-Lawson A matrix using vector magnetometer measurements.

Arguments:

Bx,By,Bz: vector magnetometer measurements [nT]

Bt: (optional) magnitude of vector magnetometer measurements or scalar magnetometer measurements for modified Tolles-Lawson [nT]

terms: (optional) Tolles-Lawson terms to use {:permanent,:induced,:eddy,:bias}

Bt_scale: (optional) scaling factor for induced and eddy current terms [nT]

Returns:

A: Tolles-Lawson A matrix

create_TL_A(flux::MagV, ind=trues(length(flux.x));
            Bt       = sqrt.(flux.x.^2+flux.y.^2+flux.z.^2)[ind],
            terms    = [:permanent,:induced,:eddy],
            Bt_scale = 50000)

Create Tolles-Lawson A matrix using vector magnetometer measurements.

Arguments:

flux: MagV vector magnetometer measurement struct

ind: (optional) selected data indices

Bt: (optional) magnitude of vector magnetometer measurements or scalar magnetometer measurements for modified Tolles-Lawson [nT]

terms: (optional) Tolles-Lawson terms to use {:permanent,:induced,:eddy,:bias}

Bt_scale: (optional) scaling factor for induced and eddy current terms [nT]

Returns:

A: Tolles-Lawson A matrix

MagNav.create_TL_coef — Function

create_TL_coef(Bx, By, Bz, B;
               Bt         = sqrt.(Bx.^2+By.^2+Bz.^2),
               λ          = 0,
               terms      = [:permanent,:induced,:eddy],
               pass1      = 0.1,
               pass2      = 0.9,
               fs         = 10.0,
               pole::Int  = 4,
               trim::Int  = 20,
               Bt_scale   = 50000,
               return_var = false)

Create Tolles-Lawson coefficients using vector and scalar magnetometer measurements and a bandpass, low-pass or high-pass filter.

Arguments:

Bx,By,Bz: vector magnetometer measurements [nT]

B: scalar magnetometer measurements [nT]

Bt: (optional) magnitude of vector magnetometer measurements or scalar magnetometer measurements for modified Tolles-Lawson [nT]

λ: (optional) ridge parameter

terms: (optional) Tolles-Lawson terms to use {:permanent,:induced,:eddy,:bias}

pass1: (optional) first passband frequency [Hz]

pass2: (optional) second passband frequency [Hz]

fs: (optional) sampling frequency [Hz]

pole: (optional) number of poles for Butterworth filter

trim (optional) number of elements to trim after filtering

Bt_scale: (optional) scaling factor for induced and eddy current terms [nT]

return_var: (optional) return B_var fit error variance along with coef

Returns:

coef: Tolles-Lawson coefficients

B_var: if return_var=true, also return fit error variance

create_TL_coef(flux::MagV, B, ind=trues(length(flux.x));
               Bt         = sqrt.(flux.x.^2+flux.y.^2+flux.z.^2)[ind],
               λ          = 0,
               terms      = [:permanent,:induced,:eddy],
               pass1      = 0.1,
               pass2      = 0.9,
               fs         = 10.0,
               pole::Int  = 4,
               trim::Int  = 20,
               Bt_scale   = 50000,
               return_var = false)

Create Tolles-Lawson coefficients using vector and scalar magnetometer measurements and a bandpass, low-pass or high-pass filter.

Arguments:

flux: MagV vector magnetometer measurement struct

B: scalar magnetometer measurements [nT]

ind: (optional) selected data indices

Bt: (optional) magnitude of vector magnetometer measurements or scalar magnetometer measurements for modified Tolles-Lawson [nT]

λ: (optional) ridge parameter

terms: (optional) Tolles-Lawson terms to use {:permanent,:induced,:eddy,:bias}

pass1: (optional) first passband frequency [Hz]

pass2: (optional) second passband frequency [Hz]

fs: (optional) sampling frequency [Hz]

pole: (optional) number of poles for Butterworth filter

trim (optional) number of elements to trim after filtering

Bt_scale: (optional) scaling factor for induced and eddy current terms [nT]

return_var: (optional) return B_var fit error variance along with coef

Returns:

coef: Tolles-Lawson coefficients

B_var: fit error variance

Model Training/Fitting

MagNav.comp_train — Function

comp_train(xyz::XYZ, ind, mapS::MapS=MapS(zeros(1,1),[0.0],[0.0],0.0);
           comp_params::CompParams=NNCompParams(), silent::Bool=true)

Train an aeromagnetic compensation model.

Arguments:

xyz: XYZ flight data struct

ind: selected data indices

mapS: (optional) MapS scalar magnetic anomaly map struct, only used for y_type = :b, :c

comp_params: CompParams aeromagnetic compensation parameters struct, either:

NNCompParams: neural network-based aeromagnetic compensation parameters struct

LinCompParams: linear aeromagnetic compensation parameters struct

silent (optional) if true, no print outs

Returns:

comp_params: CompParams aeromagnetic compensation parameters struct

y: observed data

y_hat: predicted data

err: compensation error

comp_train(lines, df_line::DataFrame, df_flight::DataFrame,
           df_map::DataFrame, comp_params::CompParams=NNCompParams();
           silent::Bool=true)

Train an aeromagnetic compensation model.

Arguments:

lines: selected line number(s)

df_line: lookup table (DataFrame) of lines

df_flight: lookup table (DataFrame) of flight files

df_map: lookup table (DataFrame) of map files

comp_params: CompParams aeromagnetic compensation parameters struct, either:

NNCompParams: neural network-based aeromagnetic compensation parameters struct

LinCompParams: linear aeromagnetic compensation parameters struct

silent (optional) if true, no print outs

Returns:

comp_params: CompParams aeromagnetic compensation parameters struct

y: observed data

y_hat: predicted data

err: mean-corrected (per line) compensation error

Model Testing

MagNav.comp_test — Function

comp_test(xyz::XYZ, ind, mapS::MapS=MapS(zeros(1,1),[0.0],[0.0],0.0);
          comp_params::CompParams=NNCompParams(), silent::Bool=false)

Evaluate aeromagnetic compensation model performance.

Arguments:

xyz: XYZ flight data struct

ind: selected data indices

mapS: (optional) MapS scalar magnetic anomaly map struct, only used for y_type = :b, :c

comp_params: CompParams aeromagnetic compensation parameters struct, either:

NNCompParams: neural network-based aeromagnetic compensation parameters struct

LinCompParams: linear aeromagnetic compensation parameters struct

silent (optional) if true, no print outs

Returns:

y: observed data

y_hat: predicted data

err: compensation error

comp_test(lines, df_line::DataFrame, df_flight::DataFrame,
          df_map::DataFrame, comp_params::CompParams=NNCompParams();
          silent::Bool=false)

Evaluate aeromagnetic compensation model performance.

Arguments:

lines: selected line number(s)

df_line: lookup table (DataFrame) of lines

df_flight: lookup table (DataFrame) of flight files

df_map: lookup table (DataFrame) of map files

comp_params: CompParams aeromagnetic compensation parameters struct

comp_params: CompParams aeromagnetic compensation parameters struct, either:

NNCompParams: neural network-based aeromagnetic compensation parameters struct

LinCompParams: linear aeromagnetic compensation parameters struct

silent (optional) if true, no print outs

Returns:

y: observed data

y_hat: predicted data

err: mean-corrected (per line) compensation error