Aeromagnetic Compensation

Placeholder page for aeromagnetic compensation documentation.

Tolles-Lawson

MagNav.create_TL_A — Function

create_TL_A(Bx, By, Bz;
            Bt       = sqrt.(Bx.^2+By.^2+Bz.^2),
            terms    = [:permanent,:induced,:eddy],
            Bt_scale = 50000)

Create Tolles-Lawson A matrix using vector magnetometer measurements.

Arguments:

Bx,By,Bz: vector magnetometer measurements [nT]

Bt: (optional) magnitude of vector magnetometer measurements or scalar magnetometer measurements for modified Tolles-Lawson [nT]

terms: (optional) Tolles-Lawson terms to use {:permanent,:induced,:eddy,:bias}

Bt_scale: (optional) scaling factor for induced and eddy current terms [nT]

Returns:

A: Tolles-Lawson A matrix

create_TL_A(flux::MagV, ind=trues(length(flux.x));
            Bt       = sqrt.(flux.x.^2+flux.y.^2+flux.z.^2)[ind],
            terms    = [:permanent,:induced,:eddy],
            Bt_scale = 50000)

Create Tolles-Lawson A matrix using vector magnetometer measurements.

Arguments:

flux: MagV vector magnetometer measurement struct

ind: (optional) selected data indices

Bt: (optional) magnitude of vector magnetometer measurements or scalar magnetometer measurements for modified Tolles-Lawson [nT]

terms: (optional) Tolles-Lawson terms to use {:permanent,:induced,:eddy,:bias}

Bt_scale: (optional) scaling factor for induced and eddy current terms [nT]

Returns:

A: Tolles-Lawson A matrix

MagNav.create_TL_coef — Function

create_TL_coef(Bx, By, Bz, B;
               Bt         = sqrt.(Bx.^2+By.^2+Bz.^2),
               λ          = 0,
               terms      = [:permanent,:induced,:eddy],
               pass1      = 0.1,
               pass2      = 0.9,
               fs         = 10.0,
               pole::Int  = 4,
               trim::Int  = 20,
               Bt_scale   = 50000,
               return_var = false)

Create Tolles-Lawson coefficients using vector and scalar magnetometer measurements and a bandpass, low-pass or high-pass filter.

Arguments:

Bx,By,Bz: vector magnetometer measurements [nT]

B: scalar magnetometer measurements [nT]

Bt: (optional) magnitude of vector magnetometer measurements or scalar magnetometer measurements for modified Tolles-Lawson [nT]

λ: (optional) ridge parameter

terms: (optional) Tolles-Lawson terms to use {:permanent,:induced,:eddy,:bias}

pass1: (optional) first passband frequency [Hz]

pass2: (optional) second passband frequency [Hz]

fs: (optional) sampling frequency [Hz]

pole: (optional) number of poles for Butterworth filter

trim (optional) number of elements to trim after filtering

Bt_scale: (optional) scaling factor for induced and eddy current terms [nT]

return_var: (optional) return B_var fit error variance along with coef

Returns:

coef: Tolles-Lawson coefficients

B_var: if return_var=true, also return fit error variance

create_TL_coef(flux::MagV, B, ind=trues(length(flux.x));
               Bt         = sqrt.(flux.x.^2+flux.y.^2+flux.z.^2)[ind],
               λ          = 0,
               terms      = [:permanent,:induced,:eddy],
               pass1      = 0.1,
               pass2      = 0.9,
               fs         = 10.0,
               pole::Int  = 4,
               trim::Int  = 20,
               Bt_scale   = 50000,
               return_var = false)

Create Tolles-Lawson coefficients using vector and scalar magnetometer measurements and a bandpass, low-pass or high-pass filter.

Arguments:

flux: MagV vector magnetometer measurement struct

B: scalar magnetometer measurements [nT]

ind: (optional) selected data indices

Bt: (optional) magnitude of vector magnetometer measurements or scalar magnetometer measurements for modified Tolles-Lawson [nT]

λ: (optional) ridge parameter

terms: (optional) Tolles-Lawson terms to use {:permanent,:induced,:eddy,:bias}

pass1: (optional) first passband frequency [Hz]

pass2: (optional) second passband frequency [Hz]

fs: (optional) sampling frequency [Hz]

pole: (optional) number of poles for Butterworth filter

trim (optional) number of elements to trim after filtering

Bt_scale: (optional) scaling factor for induced and eddy current terms [nT]

return_var: (optional) return B_var fit error variance along with coef

Returns:

coef: Tolles-Lawson coefficients

B_var: fit error variance

Linear Model Parameters

MagNav.LinCompParams — Type

struct LinCompParams <: CompParams

Linear aeromagnetic compensation parameters struct. Subtype of CompParams.

To see default parameters, type LinCompParams().

General Parameters:

Parameter	Description

features_setup	list of features to include

features_no_norm	list of features to not normalize

model_type	aeromagnetic compensation model type (see below)

y_type	y target type (see below)

use_mag	scalar magnetometer to use {:mag_1_uc, etc.}, only used for y_type = :c, :d, :e

use_vec	vector magnetometer (fluxgate) to use for Tolles-Lawson A matrix {:flux_a, etc.}, not used for model_type = :elasticnet, :plsr

data_norms	Tuple of data normalizations, e.g. (A_bias,A_scale,x_bias,x_scale,y_bias,y_scale)

weights	linear model coefficients

terms	Tolles-Lawson terms to use for Tolles-Lawson A matrix (or matrices) within x matrix {:permanent,:induced,:eddy}

terms_A	Tolles-Lawson terms to use for "external" Tolles-Lawson A matrix {:permanent,:induced,:eddy,:bias}, not used for model_type = :elasticnet, :plsr

sub_diurnal	if true, subtract diurnal from scalar magnetometer measurements

sub_igrf	if true, subtract IGRF from scalar magnetometer measurements

bpf_mag	if true, bpf scalar magnetometer measurements in x matrix

norm_type_A	normalization for Tolles-Lawson A matrix, not used for model_type = :elasticnet, :plsr (see below)

norm_type_x	normalization for x matrix (see below)

norm_type_y	normalization for y target vector (see below)

model_type options:

:TL = classical Tolles-Lawson

:mod_TL = modified Tolles-Lawson

:map_TL = map-based Tolles-Lawson

:elasticnet = elastic net (ridge regression and/or Lasso)

:plsr: = partial least squares regression (PLSR)

y_type options:

:a = anomaly field #1, compensated tail stinger total field scalar magnetometer measurements

:b = anomaly field #2, interpolated magnetic anomaly map values

:c = aircraft field #1, difference between uncompensated cabin total field scalar magnetometer measurements and interpolated magnetic anomaly map values

:d = aircraft field #2, difference between uncompensated cabin and compensated tail stinger total field scalar magnetometer measurements

:e = BPF'd total field, bandpass filtered uncompensated cabin total field scalar magnetometer measurements

norm_type options:

:standardize = Z-score normalization

:normalize = min-max normalization

:scale = scale by maximum absolute value, bias = 0

:none = scale by 1, bias = 0

Linear Model-Specific Parameters:

Parameter	Description

k_plsr	number of components, only used for model_type = :plsr

λ_TL	ridge parameter, only used for model_type = :TL, :mod_TL, :map_TL

Neural Network Model Parameters

MagNav.NNCompParams — Type

struct NNCompParams <: CompParams

Neural network-based aeromagnetic compensation parameters struct. Subtype of CompParams.

To see default parameters, type NNCompParams().

General Parameters:

Parameter	Description

features_setup	list of features to include

features_no_norm	list of features to not normalize

model_type	aeromagnetic compensation model type (see below)

y_type	y target type (see below)

use_mag	scalar magnetometer to use {:mag_1_uc, etc.}, only used for y_type = :c, :d, :e

use_vec	vector magnetometer (fluxgate) to use for Tolles-Lawson A matrix {:flux_a, etc.}, not used for model_type = :m1

data_norms	Tuple of data normalizations, e.g. (A_bias,A_scale,v_scale,x_bias,x_scale,y_bias,y_scale)

weights	neural network model weights

terms	Tolles-Lawson terms to use for Tolles-Lawson A matrix (or matrices) within x matrix {:permanent,:induced,:eddy}

terms_A	Tolles-Lawson terms to use for "external" Tolles-Lawson A matrix {:permanent,:induced,:eddy,:bias}, not used for model_type = :m1

sub_diurnal	if true, subtract diurnal from scalar magnetometer measurements

sub_igrf	if true, subtract IGRF from scalar magnetometer measurements

bpf_mag	if true, bpf scalar magnetometer measurements in x matrix

norm_type_A	normalization for Tolles-Lawson A matrix, not used for model_type = :m1 (see below)

norm_type_x	normalization for x matrix (see below)

norm_type_y	normalization for y target vector (see below)

model_type options:

:m1 = standard NN

:m2a = NN determines Tolles-Lawson coefficients

:m2b = NN determines additive correction to classical Tolles-Lawson

:m2c = NN determines additive correction to classical Tolles-Lawson, Tolles-Lawson coefficients tuned as well

:m2d = NN determines additive correction to each Tolles-Lawson coefficient

y_type options:

:a = anomaly field #1, compensated tail stinger total field scalar magnetometer measurements

:b = anomaly field #2, interpolated magnetic anomaly map values

:c = aircraft field #1, difference between uncompensated cabin total field scalar magnetometer measurements and interpolated magnetic anomaly map values

:d = aircraft field #2, difference between uncompensated cabin and compensated tail stinger total field scalar magnetometer measurements

:e = BPF'd total field, bandpass filtered uncompensated cabin total field scalar magnetometer measurements

norm_type options:

:standardize = Z-score normalization

:normalize = min-max normalization

:scale = scale by maximum absolute value, bias = 0

:none = scale by 1, bias = 0

Neural Network-Based Model-Specific Parameters:

Parameter	Description

TL_coef	Tolles-Lawson coefficients, only used for model_type = :m2b, :m2d (and :m2c initial TL_coef)

η_adam	learning rate for Adam optimizer

epoch_adam	number of epochs for Adam optimizer

epoch_lbfgs	number of epochs for LBFGS optimizer

hidden	hidden layers & nodes, e.g. [8,8] for 2 hidden layers, 8 nodes each

activation	activation function (see below)

batchsize	mini-batch size

frac_train	fraction of training data used for training (remainder for validation), only used for Adam optimizer

α_sgl	Lasso (α_sgl=0) vs group Lasso (α_sgl=1) balancing parameter {0:1}

λ_sgl	sparse group Lasso parameter, typically ~1e-5 (if nonzero)

k_pca	number of components for pre-processing with PCA + whitening, -1 to ignore

drop_fi	if true, perform drop-column feature importance

drop_fi_bson	file name (without extension) to save/load drop-column feature importance data

drop_fi_csv	file name to save drop-column feature importance data

perm_fi	if true, perform permutation feature importance

perm_fi_csv	file name to save permutation feature importance data

activation options:

relu = rectified linear unit

σ = sigmoid (logistic function)

swish = self-gated

tanh = hyperbolic tan

run plot_activation() for a visual

Model Training/Fitting

MagNav.comp_train — Function

comp_train(xyz::XYZ, ind, mapS::MapS=MapS(zeros(1,1),[0.0],[0.0],0.0);
           comp_params::CompParams=NNCompParams(), silent::Bool=true)

Train an aeromagnetic compensation model.

Arguments:

xyz: XYZ flight data struct

ind: selected data indices

mapS: (optional) MapS scalar magnetic anomaly map struct, only used for y_type = :b, :c

comp_params: CompParams aeromagnetic compensation parameters struct, either:

NNCompParams: neural network-based aeromagnetic compensation parameters struct

LinCompParams: linear aeromagnetic compensation parameters struct

silent (optional) if true, no print outs

Returns:

comp_params: CompParams aeromagnetic compensation parameters struct

y: observed data

y_hat: predicted data

err: compensation error

comp_train(lines, df_line::DataFrame, df_flight::DataFrame,
           df_map::DataFrame, comp_params::CompParams=NNCompParams();
           silent::Bool=true)

Train an aeromagnetic compensation model.

Arguments:

lines: selected line number(s)

df_line: lookup table (DataFrame) of lines

df_flight: lookup table (DataFrame) of flight files

df_map: lookup table (DataFrame) of map files

comp_params: CompParams aeromagnetic compensation parameters struct, either:

NNCompParams: neural network-based aeromagnetic compensation parameters struct

LinCompParams: linear aeromagnetic compensation parameters struct

silent (optional) if true, no print outs

Returns:

comp_params: CompParams aeromagnetic compensation parameters struct

y: observed data

y_hat: predicted data

err: mean-corrected (per line) compensation error

Model Testing

MagNav.comp_test — Function

comp_test(xyz::XYZ, ind, mapS::MapS=MapS(zeros(1,1),[0.0],[0.0],0.0);
          comp_params::CompParams=NNCompParams(), silent::Bool=false)

Evaluate aeromagnetic compensation model performance.

Arguments:

xyz: XYZ flight data struct

ind: selected data indices

mapS: (optional) MapS scalar magnetic anomaly map struct, only used for y_type = :b, :c

comp_params: CompParams aeromagnetic compensation parameters struct, either:

NNCompParams: neural network-based aeromagnetic compensation parameters struct

LinCompParams: linear aeromagnetic compensation parameters struct

silent (optional) if true, no print outs

Returns:

y: observed data

y_hat: predicted data

err: compensation error

comp_test(lines, df_line::DataFrame, df_flight::DataFrame,
          df_map::DataFrame, comp_params::CompParams=NNCompParams();
          silent::Bool=false)

Evaluate aeromagnetic compensation model performance.

Arguments:

lines: selected line number(s)

df_line: lookup table (DataFrame) of lines

df_flight: lookup table (DataFrame) of flight files

df_map: lookup table (DataFrame) of map files

comp_params: CompParams aeromagnetic compensation parameters struct

comp_params: CompParams aeromagnetic compensation parameters struct, either:

NNCompParams: neural network-based aeromagnetic compensation parameters struct

LinCompParams: linear aeromagnetic compensation parameters struct

silent (optional) if true, no print outs

Returns:

y: observed data

y_hat: predicted data

err: mean-corrected (per line) compensation error